fix: remove duplicate KeycloakRealmImport causing realm conflict
The RHTPA server was crashing because the Keycloak realm import was failing
due to duplicate KeycloakRealmImport resources trying to manage the same realm:

1. charts/keycloak creates 'ztvp-realm-import' with qtodo and RHTAS users
2. charts/rhtpa-operator created 'rhtpa-user-import' also targeting 'ztvp' realm

When both tried to import the same realm, the Keycloak operator attempted to
create the same secret ('keycloak-ztvp-realm'), causing a 409 conflict error.
This prevented the realm from being created, causing RHTPA server to fail with:
  Error: HTTP 404 for https://keycloak.../realms/ztvp/.well-known/openid-configuration

Changes:
- Added rhtpa-user to keycloak chart's defaultRealm users list
- Added RHTPA_USER_PASSWORD placeholder to keycloak-realm-import template
- Added rhtpa-user-password to keycloak-users ExternalSecret
- Removed duplicate keycloak-rhtpa-user.yaml from rhtpa-operator chart

Now there is only ONE KeycloakRealmImport that creates the complete ztvp realm
with all users (qtodo, RHTAS, RHTPA) from a single source of truth.

Fixes: RHTPA server CrashLoopBackOff due to missing Keycloak realm
Signed-off-by: Min Zhang <minzhang@redhat.com>